+++ packages/cohere/src/cohere-chat-language-model.test.ts
@@ -5,6 +5,7 @@ import {
   convertReadableStreamToArray,
 } from '@ai-sdk/provider-utils/test';
 import { createCohere } from './cohere-provider';
+import { fail } from 'assert';
 
 const TEST_PROMPT: LanguageModelV1Prompt = [
   {
@@ -14,7 +15,14 @@ const TEST_PROMPT: LanguageModelV1Prompt = [
   { role: 'user', content: [{ type: 'text', text: 'Hello' }] },
 ];
 
-const provider = createCohere({ apiKey: 'test-api-key' });
+let testIdCounter = 0;
+
+const provider = createCohere({
+  apiKey: 'test-api-key',
+  generateId: () => {
+    return `test-id-${testIdCounter++}`;
+  },
+});
 const model = provider('command-r-plus');
 
 describe('doGenerate', () => {
@@ -25,6 +33,7 @@ describe('doGenerate', () => {
   function prepareJsonResponse({
     input = '',
     text = '',
+    tool_calls,
     finish_reason = 'COMPLETE',
     tokens = {
       input_tokens: 4,
@@ -33,6 +42,7 @@ describe('doGenerate', () => {
   }: {
     input?: string;
     text?: string;
+    tool_calls?: any;
     finish_reason?: string;
     tokens?: {
       input_tokens: number;
@@ -47,6 +57,7 @@ describe('doGenerate', () => {
         { role: 'USER', message: input },
         { role: 'CHATBOT', message: text },
       ],
+      ...(tool_calls ? { tool_calls } : {}),
       finish_reason,
       meta: {
         api_version: { version: '1' },
@@ -66,6 +77,50 @@ describe('doGenerate', () => {
     });
 
     expect(text).toStrictEqual('Hello, World!');
+  });
+
+  it('should extract tool calls', async () => {
+    prepareJsonResponse({
+      text: 'Hello, World!',
+      tool_calls: [
+        {
+          name: 'test-tool',
+          parameters: { value: 'example value' },
+        },
+      ],
+    });
+
+    const { text, toolCalls, finishReason } = await model.doGenerate({
+      inputFormat: 'prompt',
+      mode: {
+        type: 'regular',
+        tools: [
+          {
+            type: 'function',
+            name: 'test-tool',
+            parameters: {
+              type: 'object',
+              properties: { value: { type: 'string' } },
+              required: ['value'],
+              additionalProperties: false,
+              $schema: 'http://json-schema.org/draft-07/schema#',
+            },
+          },
+        ],
+      },
+      prompt: TEST_PROMPT,
+    });
+
+    expect(toolCalls).toStrictEqual([
+      {
+        toolCallId: expect.any(String),
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        args: '{"value":"example value"}',
+      },
+    ]);
+    expect(text).toStrictEqual('Hello, World!');
+    expect(finishReason).toStrictEqual('stop');
   });
 
   it('should extract usage', async () => {
@@ -135,6 +190,155 @@ describe('doGenerate', () => {
       model: 'command-r-plus',
       message: 'Hello',
       chat_history: [{ role: 'SYSTEM', message: 'you are a friendly bot!' }],
+    });
+  });
+
+  describe('should pass tools', async () => {
+    it('should convert primitive types', async () => {
+      prepareJsonResponse({});
+
+      await model.doGenerate({
+        inputFormat: 'prompt',
+        mode: {
+          type: 'regular',
+          tools: [
+            {
+              type: 'function',
+              name: 'test-tool',
+              parameters: {
+                type: 'object',
+                properties: {
+                  value_a: { type: 'string' },
+                  value_b: { type: 'number' },
+                  value_c: { type: 'integer' },
+                  value_d: { type: 'boolean' },
+                },
+                required: ['value_a'],
+                additionalProperties: false,
+                $schema: 'http://json-schema.org/draft-07/schema#',
+              },
+            },
+          ],
+        },
+        prompt: TEST_PROMPT,
+      });
+
+      expect(await server.getRequestBodyJson()).toStrictEqual({
+        model: 'command-r-plus',
+        chat_history: [
+          {
+            role: 'SYSTEM',
+            message: 'you are a friendly bot!',
+          },
+        ],
+        force_single_step: false,
+        message: 'Hello',
+        tools: [
+          {
+            name: 'test-tool',
+            parameterDefinitions: {
+              value_a: {
+                type: 'str',
+                required: true,
+              },
+              value_b: {
+                type: 'float',
+                required: false,
+              },
+              value_c: {
+                type: 'int',
+                required: false,
+              },
+              value_d: {
+                type: 'bool',
+                required: false,
+              },
+            },
+          },
+        ],
+      });
+    });
+
+    it('should throw error for unsupported types', async () => {
+      prepareJsonResponse({});
+
+      await expect(
+        model.doGenerate({
+          inputFormat: 'prompt',
+          mode: {
+            type: 'regular',
+            tools: [
+              {
+                type: 'function',
+                name: 'test-tool',
+                parameters: {
+                  type: 'object',
+                  properties: {
+                    value: { type: 'array', items: { type: 'string' } },
+                  },
+                  required: ['value'],
+                  additionalProperties: false,
+                  $schema: 'http://json-schema.org/draft-07/schema#',
+                },
+              },
+            ],
+          },
+          prompt: TEST_PROMPT,
+        }),
+      ).rejects.toThrow();
+    });
+
+    it('should pass tool choice', async () => {
+      prepareJsonResponse({});
+
+      await model.doGenerate({
+        inputFormat: 'prompt',
+        mode: {
+          type: 'regular',
+          toolChoice: {
+            type: 'required',
+          },
+          tools: [
+            {
+              type: 'function',
+              name: 'test-tool',
+              parameters: {
+                type: 'object',
+                properties: {
+                  value: { type: 'string' },
+                },
+                required: ['value'],
+                additionalProperties: false,
+                $schema: 'http://json-schema.org/draft-07/schema#',
+              },
+            },
+          ],
+        },
+        prompt: TEST_PROMPT,
+      });
+
+      expect(await server.getRequestBodyJson()).toStrictEqual({
+        model: 'command-r-plus',
+        chat_history: [
+          {
+            role: 'SYSTEM',
+            message: 'you are a friendly bot!',
+          },
+        ],
+        force_single_step: true,
+        message: 'Hello',
+        tools: [
+          {
+            name: 'test-tool',
+            parameterDefinitions: {
+              value: {
+                type: 'str',
+                required: true,
+              },
+            },
+          },
+        ],
+      });
     });
   });
 
@@ -276,6 +480,422 @@ describe('doStream', () => {
         usage: { promptTokens: 34, completionTokens: 12 },
       },
     ]);
+  });
+
+  it('should stream tool deltas', async () => {
+    server.responseChunks = [
+      `{"event_type":"stream-start"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":"I"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" will"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" use"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" the"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" get"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":"Stock"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":"Price"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" tool"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" to"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" find"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" the"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" price"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" of"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" AAPL"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":" stock"}\n\n`,
+      `{"event_type":"tool-calls-chunk","text":"."}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"name":"test-tool"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"{\\n    \\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"ticker"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"_"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"symbol"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"\\":"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":" \\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"AAPL"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"\\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"\\n"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"}"}}\n\n`,
+      `{"event_type":"tool-calls-generation","tool_calls":[{"name":"test-tool-a","parameters":{"ticker_symbol":"AAPL"}}]}\n\n`,
+      `{"event_type":"stream-end","finish_reason":"COMPLETE","response":{"meta":{"tokens":{"input_tokens":893,"output_tokens":62}}}}\n\n`,
+    ];
+
+    const { stream } = await model.doStream({
+      inputFormat: 'prompt',
+      prompt: TEST_PROMPT,
+      mode: {
+        type: 'regular',
+        tools: [
+          {
+            type: 'function',
+            name: 'test-tool',
+            parameters: {
+              type: 'object',
+              properties: { value: { type: 'string' } },
+              required: ['value'],
+              additionalProperties: false,
+              $schema: 'http://json-schema.org/draft-07/schema#',
+            },
+          },
+        ],
+      },
+    });
+
+    const responseArray = await convertReadableStreamToArray(stream);
+
+    expect(responseArray).toStrictEqual([
+      {
+        type: 'tool-call-delta',
+        toolCallType: 'function',
+        toolCallId: 'test-id-1',
+        toolName: 'test-tool',
+        argsTextDelta: '',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: '{\n    "',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: 'ticker',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: '_',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: 'symbol',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: '":',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: ' "',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: 'AAPL',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: '"',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: '\n',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        argsTextDelta: '}',
+      },
+      {
+        type: 'tool-call',
+        toolCallId: 'test-id-1',
+        toolCallType: 'function',
+        toolName: 'test-tool',
+        args: '{"ticker_symbol":"AAPL"}',
+      },
+      {
+        finishReason: 'stop',
+        type: 'finish',
+        usage: {
+          completionTokens: 62,
+          promptTokens: 893,
+        },
+      },
+    ]);
+
+    // Check if the tool call ID is the same in the tool call delta and the tool call
+    const toolCallIds = responseArray
+      .filter(
+        chunk => chunk.type === 'tool-call-delta' || chunk.type === 'tool-call',
+      )
+      .map(chunk => chunk.toolCallId);
+
+    expect(new Set(toolCallIds)).toStrictEqual(new Set(['test-id-1']));
+  });
+
+  it('should handle out of order tool deltas', async () => {
+    server.responseChunks = [
+      `{"event_type":"stream-start"}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"name":"test-tool-a"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"name":"test-tool-b"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"{\\n    \\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"{\\n    \\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"ticker"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"ticker"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"_"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"_"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"symbol"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"symbol"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"\\":"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"\\":"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":" \\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":" \\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"TSLA"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"AAPL"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"\\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"\\""}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"\\n"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"\\n"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":0,"parameters":"}"}}\n\n`,
+      `{"event_type":"tool-calls-chunk","tool_call_delta":{"index":1,"parameters":"}"}}\n\n`,
+      `{"event_type":"tool-calls-generation","tool_calls":[{"name":"test-tool-a","parameters":{"ticker_symbol":"AAPL"}},{"name":"test-tool-b","parameters":{"ticker_symbol":"TSLA"}}]}\n\n`,
+      `{"event_type":"stream-end","finish_reason":"COMPLETE","response":{"meta":{"tokens":{"input_tokens":893,"output_tokens":62}}}}\n\n`,
+    ];
+
+    const { stream } = await model.doStream({
+      inputFormat: 'prompt',
+      prompt: TEST_PROMPT,
+      mode: {
+        type: 'regular',
+        tools: [
+          {
+            type: 'function',
+            name: 'test-tool-a',
+            parameters: {
+              type: 'object',
+              properties: { value: { type: 'string' } },
+              required: ['value'],
+              additionalProperties: false,
+              $schema: 'http://json-schema.org/draft-07/schema#',
+            },
+          },
+          {
+            type: 'function',
+            name: 'test-tool-b',
+            parameters: {
+              type: 'object',
+              properties: { value: { type: 'string' } },
+              required: ['value'],
+              additionalProperties: false,
+              $schema: 'http://json-schema.org/draft-07/schema#',
+            },
+          },
+        ],
+      },
+    });
+
+    const responseArray = await convertReadableStreamToArray(stream);
+
+    expect(responseArray).toStrictEqual([
+      {
+        type: 'tool-call-delta',
+        toolCallType: 'function',
+        toolCallId: 'test-id-2',
+        toolName: 'test-tool-a',
+        argsTextDelta: '',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallType: 'function',
+        toolCallId: 'test-id-3',
+        toolName: 'test-tool-b',
+        argsTextDelta: '',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: '{\n    "',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: '{\n    "',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: 'ticker',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: 'ticker',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: '_',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: '_',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: 'symbol',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: 'symbol',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: '":',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: '":',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: ' "',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: ' "',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: 'TSLA',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: 'AAPL',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: '"',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: '"',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: '\n',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: '\n',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        argsTextDelta: '}',
+      },
+      {
+        type: 'tool-call-delta',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        argsTextDelta: '}',
+      },
+      {
+        type: 'tool-call',
+        toolCallId: 'test-id-2',
+        toolCallType: 'function',
+        toolName: 'test-tool-a',
+        args: '{"ticker_symbol":"AAPL"}',
+      },
+      {
+        type: 'tool-call',
+        toolCallId: 'test-id-3',
+        toolCallType: 'function',
+        toolName: 'test-tool-b',
+        args: '{"ticker_symbol":"TSLA"}',
+      },
+      {
+        finishReason: 'stop',
+        type: 'finish',
+        usage: {
+          completionTokens: 62,
+          promptTokens: 893,
+        },
+      },
+    ]);
+
+    // Check if the tool call ID is the same in the tool call delta and the tool call
+    const toolCallIds = responseArray
+      .filter(
+        chunk => chunk.type === 'tool-call-delta' || chunk.type === 'tool-call',
+      )
+      .map(chunk => chunk.toolCallId);
+
+    expect(new Set(toolCallIds)).toStrictEqual(
+      new Set(['test-id-2', 'test-id-3']),
+    );
   });
 
   it('should handle unparsable stream parts', async () => {